fixed compiler issue in which gcc-3.4.3 rejected to convert an int to an unsigned int

+++ Packages/QP_solver/include/CGAL/QP_solver/QP__partial_base.h
@@ -125,7 +125,8 @@ init( )
     int  n = this->solver().number_of_variables();
     int  m = this->solver().number_of_constraints();
 
-    s = std::min( static_cast< unsigned int>( m*std::sqrt( n/2.0)), N.size());
+    s = std::min( static_cast< unsigned int>( m*std::sqrt( n/2.0)),
+		  static_cast< unsigned int>(N.size()));
 }
 
 // operations
@@ -193,7 +194,8 @@ transition( )
     int  n = this->solver().number_of_variables();
     int  m = this->solver().number_of_constraints();
 
-    s = std::min( static_cast< unsigned int>( m*std::sqrt( n/2.0)), N.size());
+    s = std::min( static_cast< unsigned int>( m*std::sqrt( n/2.0)),
+		  static_cast< unsigned int>(N.size()));
 }
 
 CGAL_END_NAMESPACE
